requirements1.5.txt
Django==1.5.12
PyCrypto
pytest
pytest-django
pytest-pep8